+++ src/main/java/io/github/astrapi69/swing/component/JMSpinner.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.swing;
+package io.github.astrapi69.swing.component;
 
 import javax.swing.*;
 import javax.swing.text.DefaultFormatter;