+++ src/main/java/io/github/astrapi69/swing/component/JMTextArea.java
@@ -1,0 +1,113 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.swing.component;
+
+import javax.swing.*;
+import javax.swing.event.DocumentEvent;
+import javax.swing.text.Document;
+
+import lombok.AccessLevel;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NonNull;
+import lombok.experimental.FieldDefaults;
+import io.github.astrapi69.model.api.Model;
+import io.github.astrapi69.swing.listener.document.DocumentListenerAdapter;
+import io.github.astrapi69.throwable.RuntimeExceptionDecorator;
+
+@Getter
+@EqualsAndHashCode(callSuper = true)
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class JMTextArea extends JTextArea
+{
+
+	/** The model. */
+	Model<String> propertyModel;
+
+	{
+		getDocument().addDocumentListener(new DocumentListenerAdapter()
+		{
+			@Override
+			public void onDocumentChanged(final DocumentEvent documentEvent)
+			{
+				int currentLength = documentEvent.getDocument().getLength();
+				final String text = RuntimeExceptionDecorator
+					.decorate(() -> documentEvent.getDocument().getText(0, currentLength));
+				if (JMTextArea.this.propertyModel != null)
+				{
+					JMTextArea.this.propertyModel.setObject(text);
+				}
+			}
+		});
+	}
+
+	/**
+	 * Constructs a new <code>TextField</code>.
+	 *
+	 * @param propertyModel
+	 *            the text model to be displayed
+	 */
+	public JMTextArea(final @NonNull Model<String> propertyModel)
+	{
+		this.propertyModel = propertyModel;
+	}
+
+	public JMTextArea(String text, Model<String> propertyModel)
+	{
+		super(text);
+		this.propertyModel = propertyModel;
+	}
+
+	public JMTextArea(int rows, int columns, Model<String> propertyModel)
+	{
+		super(rows, columns);
+		this.propertyModel = propertyModel;
+	}
+
+	public JMTextArea(String text, int rows, int columns, Model<String> propertyModel)
+	{
+		super(text, rows, columns);
+		this.propertyModel = propertyModel;
+	}
+
+	public JMTextArea(Document doc, Model<String> propertyModel)
+	{
+		super(doc);
+		this.propertyModel = propertyModel;
+	}
+
+	public JMTextArea(Document doc, String text, int rows, int columns, Model<String> propertyModel)
+	{
+		super(doc, text, rows, columns);
+		this.propertyModel = propertyModel;
+	}
+
+	public JMTextArea setPropertyModel(final @NonNull Model<String> propertyModel)
+	{
+		this.propertyModel = propertyModel;
+		setText(this.propertyModel.getObject());
+		return this;
+	}
+}